move api-service into src (whoops)
